requirements.txt
apache-airflow==2.8.4
python-dotenv==1.0.1

google-cloud-bigquery==3.40.0
google-auth==2.48.0
